+++ server/gateway/DatapackDownloader_sub.cpp
@@ -25,7 +25,11 @@ void DatapackDownloaderMainSub::writeNewFileSub(const QString &fileName,const QB
         QFile file(fullPath);
         QFileInfo fileInfo(file);
 
-        QDir(fileInfo.absolutePath()).mkpath(fileInfo.absolutePath());
+        if(!QDir(fileInfo.absolutePath()).mkpath(fileInfo.absolutePath()))
+        {
+            qDebug() << "unable to make the path: " << fileInfo.absolutePath();
+            abort();
+        }
 
         if(file.exists())
             if(!file.remove())
@@ -61,10 +65,23 @@ bool DatapackDownloaderMainSub::getHttpFileSub(const QString &url, const QString
     if(!httpModeSub)
         httpModeSub=true;
 
-    FILE *fp = fopen(fileName.toLocal8Bit().constData(),"wb");
+    const QString &fullPath=mDatapackSub+text_slash+fileName;
+    {
+        QFile file(fullPath);
+        QFileInfo fileInfo(file);
+
+        if(!QDir(fileInfo.absolutePath()).mkpath(fileInfo.absolutePath()))
+        {
+            qDebug() << "unable to make the path: " << fileInfo.absolutePath();
+            abort();
+        }
+    }
+
+    FILE *fp = fopen(fullPath.toLocal8Bit().constData(),"wb");
     if(fp!=NULL)
     {
         curl_easy_setopt(curl, CURLOPT_URL, url.toUtf8().constData());
+        curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, &fwrite);
         curl_easy_setopt(curl, CURLOPT_WRITEDATA, fp);
         const CURLcode res = curl_easy_perform(curl);
         long http_code = 0;
@@ -429,18 +446,27 @@ void DatapackDownloaderMainSub::httpFinishedForDatapackListSub(const QByteArray 
                             if(!file.exists())
                             {
                                 if(!getHttpFileSub(selectedMirror+fileString,fileString))
+                                {
+                                    datapackDownloadError();
                                     return;
+                                }
                             }
                             else if(hashFileOnDisk!=*reinterpret_cast<const quint32 *>(QByteArray::fromHex(partialHashString.toLatin1()).constData()))
                             {
                                 if(!getHttpFileSub(selectedMirror+fileString,fileString))
+                                {
+                                    datapackDownloadError();
                                     return;
+                                }
                             }
                         }
                         else
                         {
                             if(!getHttpFileSub(selectedMirror+fileString,fileString))
+                            {
+                                datapackDownloadError();
                                 return;
+                            }
                         }
                         partialHashListSub.removeAt(indexInDatapackList);
                         datapackFilesListSub.removeAt(indexInDatapackList);
